requirements.txt
amqp==5.1.1
aniso8601==9.0.1
asgiref==3.7.2
autopep8==2.0.4
billiard==4.1.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.0
cryptography==41.0.3
defusedxml==0.7.1
Deprecated==1.2.14
dj-rest-auth==4.0.1
Django==4.2.4
django-allauth==0.55.2
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-cors-headers==4.2.0
django-environ==0.11.1
django-filter==23.2
django-oauth-toolkit==2.3.0
django-rest-framework==0.1.0
django-restframework==0.0.1
django-timezone-field==6.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-social-oauth2==2.1.3
ecdsa==0.18.0
environ==1.0
graphene==3.3
graphene-django==3.1.5
graphql-core==3.2.3
graphql-relay==3.2.0
idna==3.4
jwcrypto==1.5.0
kombu==5.3.2
Markdown==3.4.4
oauthlib==3.2.2
postgres==4.0
promise==2.3
prompt-toolkit==3.0.39
psycopg2-binary==2.9.7
psycopg2-pool==1.1
pyasn1==0.5.0
pycodestyle==2.11.0
pycparser==2.21
PyJWT==2.8.0
python-crontab==3.0.0
python-dateutil==2.8.2
python-jose==3.3.0
python3-openid==3.2.0
pytz==2023.3
redis==5.0.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
social-auth-app-django==5.3.0
social-auth-core==4.4.2
sqlparse==0.4.4
text-unidecode==1.3
tzdata==2023.3
urllib3==2.0.4
vine==5.0.0
wcwidth==0.2.6
wrapt==1.15.0
